% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cispr_limits.R
\name{compute_en55011}
\alias{compute_en55011}
\title{Compute the limits for given frequencies}
\usage{
compute_en55011(frequency, type = "AVG", class = "B")
}
\arguments{
\item{frequency}{numeric vector of frequencies}

\item{type}{character, type of the detector: average or Quasi-Peak}

\item{class}{character, class of the device "A" or "B"}
}
\value{
numeric vector, limits for
}
\description{
Compute the limits in dBuV as prescribed by the standards EN55011, for a given
class of device (A or B) and type of spectrum (Average or Quasi-Peak).
}
\details{
For the frequencies outside the limits, this function returns NA_real_.
}
