% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cispr_limits.R
\name{get_limits}
\alias{get_limits}
\title{Get the standardised limits according to EN55011}
\usage{
get_limits(type = "Avg", class = "B")
}
\arguments{
\item{type}{"Avg" (Average) or "QP" (Quasi-Peak)}

\item{class}{"A" or "B"}
}
\value{

}
\description{
Get the standardised limits according to EN55011
}
