% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cispr_limits.R
\name{limits_fun}
\alias{limits_fun}
\alias{log_limits_fn}
\alias{const_limits_fn}
\alias{select_fn}
\title{Compute limits}
\usage{
log_limits_fn(start_freq, end_freq, start_val, end_val)

const_limits_fn(start_freq, end_freq, start_val, end_val = start_val)

select_fn(type, start_freq, end_freq, start_val, end_val)
}
\arguments{
\item{start_freq}{numeric, start frequency}

\item{end_freq}{numeric, end frequency}

\item{start_val}{numeric, starting limit}

\item{end_val}{numeric, ending limit}

\item{type}{character, the type of the limit ("log" or "const")}
}
\value{
function to compute the limits given the frequency
}
\description{
\code{log_limits_fn} creates the logarithmic limits, \code{const_limits_fn} computes
constant limits, \code{select_fn} select appropriate function based on the \code{type}
argument.
}
\keyword{internal}
