% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_receiver_scan_txt}
\alias{read_receiver_scan_txt}
\title{Read scans}
\usage{
read_receiver_scan_txt(full_path, join = FALSE)
}
\arguments{
\item{full_path}{a path to the receiver scan file}

\item{join}{logical (default FALSE), should the header and data be joined?}
}
\value{
tibble or a list of tibbles
}
\description{
\code{read_receiver_scan_txt} reads the scans saved as text files.
}
\details{
Header and data table are combined by the common `Scan Number`. If \code{join} is
\code{TRUE}, the header tibble and data tibble are joined to form a one large tibble with all the metadata.
Otherwise (default) the function returns a list with two fields: `header` and `data`. The common columns is
the `Scan Number`.
}
